\name{pos}
\alias{pos}
\docType{data}
\title{
Sample matrix
}
\description{
The CpG positions.
It is part of the sample dataset exampleHummingbird.
}
\usage{pos}
\source{
Chen et al. (2017) Global misregulation of genes largely uncoupled to DNA
methylome epimutations characterizes a congenital overgrowth syndrome.
Scientific Reports 7, 12667

The raw FASTQ files of the WGBS experiment from this study are publicly
available at Gene Expression Omnibus (GEO) database with accession no. GSE93775.
}
\examples{
data(exampleHummingbird)
pos
}
\keyword{datasets}
