\name{exampleHummingbird}
\alias{exampleHummingbird}
\docType{data}
\title{
Sample dataset
}
\description{
Example of input and output data of the hummingbird package. Data used as input to the hummingbird package: m_normM, m_normUM, m_abnormM, m_abnormUM, m_pos. Output data generated by the hummingbird package: hmmbird1, hmmbird2.
}
\usage{data("exampleHummingbird")}
\format{
    Matrices.
    \describe{
    \item{\code{m_normM}}{
A matrix containing the methylated read count data of the normal group. Each column of the matrix represents a replicate and each row represents a CpG position.
}
    \item{\code{m_normUM}}{
A matrix containing the unmethylated read count data of the normal group. Each column of the matrix represents a replicate and each row represents a CpG position.
}
    \item{\code{m_abnormM}}{
A matrix containing the methylated read count data of the abnormal group. Each column of the matrix represents a replicate and each row represents a CpG position.
}
    \item{\code{m_abnormUM}}{
A matrix containing the unmethylated read count data of abnormal group. Each column of the matrix represents a replicate and each row represents a CpG position.
}
    \item{\code{m_pos}}{
The CpG positions.
}
    \item{\code{hmmbird1}}{
The output of the hummingbirdEM function.
}
    \item{\code{hmmbird2}}{
The output of the hummingbirdPostAdjustment function.
}
}
}
\source{
The raw FASTQ files are publicly available at Gene Expression Omnibus, GEO accession no. GSE93775.
Chen, Z., Hagen, D. E., Ji, T., Elsik, C. G., and Rivera, R. M. (2017). Global misregulation of genes largely uncoupled to DNA methylome epimutations characterizes a congenital overgrowth syndrome. Scientific Reports 7, 12667.
}
\examples{
data(exampleHummingbird)
str(hmmbird1)
str(hmmbird2)
str(m_abnormM)
str(m_abnormUM)
str(m_normM)
str(m_normUM)
str(m_pos)
}
\keyword{datasets}
