\name{hummingbird-package}
\alias{hummingbird-package}
\alias{hummingbird}
\docType{package}
\title{
    A Bayesian Hidden Markov Model for the detection of differentially methylated regions
}
\description{
    The package detects differentially methylated regions between case and control groups using whole genome bisulfite sequencing or reduced representative bisulfite sequencing experiment data.

    The hummingbird package implements the algorithm described in the publication below.
}
\details{
    The main functions of the package are: hummingbirdEM, hummingbirdPostAdjustment and hummingbirdGraph.
}
\author{
Eleni Adam

Maintainer: Eleni Adam <eadam002@odu.edu>
}
\references{
    Ji T. A Bayesian hidden Markov model for detecting differentially methylated regions. Biometrics. 2019;75(2):663-673. DOI:10.1111/biom.13000
}
